+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The structure and content of this file follows [Keep a Changelog](https://keepac
 - Added support for root fragments in filters such as `$.data[?(@.id == $.key)]`.
 - "exists" is now an alias for the "has" filter operation.
 - Added length, count, match, and search functions.
+- Added `Nothing` as a value for comparison to return values where nothing is found.
 
 - TBD support [?foo